requirements.txt
streamlit
pandas
google-generativeai
pydantic
google-search-results